+++ src/components/anchor/anchor.jsx
@@ -2,6 +2,7 @@
 
 const React = require('react')
 const DataAttributesMixin = require('react-data-attributes-mixin')
+const PropTypes = require('prop-types')
 
 module.exports = React.createClass({
 
@@ -15,13 +16,13 @@ module.exports = React.createClass({
   },
 
   propTypes: {
-    handleClick: React.PropTypes.func,
-    children: React.PropTypes.node,
-    data: React.PropTypes.object,
-    href: React.PropTypes.string,
-    target: React.PropTypes.oneOf(['_self', '_blank', '_parent', '_top']),
-    title: React.PropTypes.string,
-    role: React.PropTypes.string
+    handleClick: PropTypes.func,
+    children: PropTypes.node,
+    data: PropTypes.object,
+    href: PropTypes.string,
+    target: PropTypes.oneOf(['_self', '_blank', '_parent', '_top']),
+    title: PropTypes.string,
+    role: PropTypes.string
   },
 
   render: function () {