+++ src/components/countdown/countdown.jsx
@@ -3,13 +3,14 @@
 const React = require('react')
 const CountdownManager = require('../../../src/components/countdown/lib/countdownManager')
 const classNames = require('classnames')
+const PropTypes = require('prop-types')
 
 module.exports = React.createClass({
 
   propTypes: {
-    purpose: React.PropTypes.oneOf(['default', 'primary', 'secondary', 'success', 'warning', 'danger', 'info']),
-    size: React.PropTypes.oneOf(['default', 'small', 'medium', 'large', 'extra-large']),
-    until: React.PropTypes.string
+    purpose: PropTypes.oneOf(['default', 'primary', 'secondary', 'success', 'warning', 'danger', 'info']),
+    size: PropTypes.oneOf(['default', 'small', 'medium', 'large', 'extra-large']),
+    until: PropTypes.string
   },
 
   getInitialState: function () {