+++ src/components/list-term/list-term.jsx
@@ -1,11 +1,12 @@
 'use strict'
 
 const React = require('react')
+const PropTypes = require('prop-types')
 
 module.exports = React.createClass({
 
   propTypes: {
-    children: React.PropTypes.any
+    children: PropTypes.any
   },
 
   render: function () {